refactor(test): check FilterMaster deserializer
